Rename: HistoryBackButton->  BackToMainButton
