--- client/src/components/HistoryBackButton.tsx
@@ -1,24 +1,0 @@
-import { useNavigate } from 'react-router-dom'
-
-const HistoryBackButton = () => {
-  const navigate = useNavigate()
-  const handleClickHistoryBackButton = () => {
-    navigate(-1)
-  }
-  return (
-    <button onClick={handleClickHistoryBackButton}>
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        stroke-width="1.5"
-        stroke="currentColor"
-        className="h-12 w-12 hover:stroke-blue-500"
-      >
-        <path stroke-linecap="round" stroke-linejoin="round" d="M9 15L3 9m0 0l6-6M3 9h12a6 6 0 010 12h-3" />
-      </svg>
-    </button>
-  )
-}
-
-export default HistoryBackButton
